requirements.txt
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
importlib==1.0.4
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
multitasking==0.0.11
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tenacity==8.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.41
